LC-find-the-smallest-divisor-given-a-threshold [Runtime: N/A; Memory: N/A]

+++ array/find-the-smallest-divisor-given-a-threshold.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    bool possible(vector<int>&nums,int mid,int threshold){
+    bool possible(vector<int>&nums,long long mid,long long threshold){
         long long cnt = 0;
         for( int it : nums){
             cnt += (it+mid-1)/mid;
@@ -8,11 +8,11 @@ public:
         if (cnt <= threshold) return true;
         else return false;
     }
-    int smallestDivisor(vector<int>& nums, int threshold) {
+    int smallestDivisor(vector<int>& nums, long long threshold) {
         int low = 0;
         int high = *max_element(nums.begin(),nums.end());
         while(low <= high){
-            int mid = low + (high - low)/2;
+            long long mid = low + (high - low)/2;
             if(possible(nums,mid,threshold)){
                 high = mid - 1;
             }
